Move getTimestampFromDateString from filter-input to time utils.
