+++ frontend/src/components/Utils/filter-input.js
@@ -13,22 +13,6 @@ export function getCleanedTextInput(rawInput, defaultValue) {
 }
 
 /**
- * Return a Firestore Timestamp corresponding to the date in `dateStr`.
- *
- * @param {string} dateStr String containing a date in the form 'yyyy-mm-dd'.
- * @return {firebase.firestore.Timestamp} Firestore timestamp object created.
- */
-export function getTimestampFromDateString(dateStr) {
-  const dateParts = dateStr.split('-').map(str => +str);
-  if (dateParts.length === 1 && dateParts[0] === 0) {
-    return firebase.firestore.Timestamp.now();
-  }
-
-  const date = new Date(dateParts[0], dateParts[1] - 1, dateParts[2]);
-  return firebase.firestore.Timestamp.fromDate(date);
-}
-
-/**
  * Return an array of collaborator uids given the emails provided in the
  * add trip form.
  *
